Move TrainingEventPalette to dto, fixing the service-to-controller layering inversion
TrainingStatsServiceImpl imported TrainingEventPalette from the controller package,
the first service->controller dependency in the codebase and a reversal of the
documented controller -> service -> repository -> model layering. The palette is
presentation data with no controller dependency of its own, so it belongs in dto.
Pure move: package declaration and every import updated, no content changes.
